refactor: 숫자들의 검증 규칙을 상위 추상 클래스로 분리하여 상속하도록 변경
