+++ src/main/java/lotto/domain/vo/lotto/Lotto.java
@@ -1,43 +1,15 @@
 package lotto.domain.vo.lotto;
 
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
-import lotto.exception.LottoError;
+import lotto.domain.rule.LotteryPolicy;
 
-public class Lotto {
-
-    private static final int FIXED_COUNT = 6;
-    private static final int MINIMUM_BOUND = 1;
-    private static final int MAXIMUM_BOUND = 45;
+public class Lotto extends LotteryPolicy {
 
     private final List<Integer> numbers;
 
     public Lotto(List<Integer> numbers) {
-        validateCountOf(numbers);
-        validateUnique(numbers);
-        numbers.forEach(this::validateRange);
+        validateAll(numbers);
         this.numbers = numbers;
-    }
-
-    private void validateCountOf(List<Integer> numbers) {
-        if (numbers.size() != FIXED_COUNT) {
-            throw new IllegalArgumentException(LottoError.NUMBER_COUNTS_ARE_NOT_MATCH.messageOf(FIXED_COUNT));
-        }
-    }
-
-    private void validateUnique(List<Integer> numbers) {
-        Set<Integer> uniqueNumbers = new HashSet<>(numbers);
-
-        if (uniqueNumbers.size() != numbers.size()) {
-            throw new IllegalArgumentException(LottoError.NUMBERS_ARE_NOT_UNIQUE.message());
-        }
-    }
-
-    private void validateRange(int number) {
-        if (number < MINIMUM_BOUND || number > MAXIMUM_BOUND) {
-            throw new IllegalArgumentException(LottoError.OUT_OF_BOUNDS.messageOf(MINIMUM_BOUND, MAXIMUM_BOUND));
-        }
     }
 
     public List<Integer> sortByAscending() {
